bug with addition of data-types
When the function `requestAuthorization` is invoked multiple times with a list of types containing the type `BODY_MASS_INDEX`, the function tends to incrementally add `HEIGHT` and `WEIGHT` on every invocation. 
The suggested changes will only add the data types if they are not already present in the input list.